Minor hacks to make the code build (at least with --debug) on Cygwin in Microsoft Windows (unknown version of GCC).
git-svn-id: https://gxemul.svn.sourceforge.net/svnroot/gxemul/gxemul/trunk@5526 dbbdf1f2-f3a0-11dc-8807-f0e41a653009

+++ src/components/cpu/M88K_CPUComponent.cc
@@ -159,7 +159,7 @@ size_t M88K_CPUComponent::DisassembleInstruction(uint64_t vaddr, size_t maxLen,
 
 	// ... and add it to the result:
 	char tmp[9];
-	snprintf(tmp, sizeof(tmp), "%08x", iword);
+	snprintf(tmp, sizeof(tmp), "%08x", (int) iword);
 	result.push_back(tmp);
 
 	int hi6 = iword >> 26;
